BED-5717 chore: moved the bheClient variable to a structure of its own
