+++ client/bloodhound/response.go
@@ -1,0 +1,5 @@
+package bloodhound
+
+type bloodhoundResponse[T any] struct {
+	Data T `json:"data"`
+}
